% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{add_ids}
\alias{add_ids}
\title{Add unique identifiers to each dataset in a list of datasets}
\usage{
add_ids(data_list)
}
\arguments{
\item{data_list}{A list of named datasets.}
}
\value{
A list of the same datasets but with an additional column set_id that serves
as row ID's.
}
\description{
Add unique identifiers to each dataset in a list of datasets
}
\keyword{internal}
