% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{calc_x_range}
\alias{calc_x_range}
\title{Calculate range of dates/study days that fall into cursor position +/- 5 px (x-axis)}
\usage{
calc_x_range(hover)
}
\arguments{
\item{hover}{Hover event of the main plot.}
}
\value{
Numeric vector of length 2.
}
\description{
Calculate range of dates/study days that fall into cursor position +/- 5 px (x-axis)
}
\keyword{internal}
