% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{check_date_type}
\alias{check_date_type}
\title{Check if the specified columns are of type Date}
\usage{
check_date_type(df, var_names)
}
\arguments{
\item{df}{A data frame.}

\item{var_names}{A string vector of column names to be checked.}
}
\value{
If all variables are of type Date, the function returns
the data frame invisibly.
}
\description{
\code{check_date_type} produces an error if data in the variables is not of type
Date.
}
\keyword{internal}
