% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_local_filter.R
\name{check_filters}
\alias{check_filters}
\title{Check if the user specified adverse event filters are available}
\usage{
check_filters(filter_list, filter_names)
}
\arguments{
\item{filter_list}{A string vector of filter names specified by the user.}

\item{filter_names}{A string vector of available filters.}
}
\description{
\code{check_filters} produces an error if the filter names can't be found
among the available local adverse event filters.
}
\keyword{internal}
