% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{check_names}
\alias{check_names}
\title{Check if the user specified variable names exist in the dataset}
\usage{
check_names(df, var_names, subjid_var)
}
\arguments{
\item{df}{A data frame.}

\item{var_names}{A character vector of column names to be checked.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
If all variables were found in the data frame, the function returns
the data frame invisibly.
}
\description{
\code{check_names} produces an error if the variables can't be found in the
data frame.
}
\keyword{internal}
