% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{check_valid_color}
\alias{check_valid_color}
\title{Check if colors are valid}
\usage{
check_valid_color(color_palette)
}
\arguments{
\item{color_palette}{Vector containing strings that are meant to specify
a color, either as 6-digit hex color starting with the # symbol, or as
R color string like received when executing \code{colors()}.}
}
\description{
Executes an error action in case of wrong color definitions.
}
\keyword{internal}
