% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{collapse_cols}
\alias{collapse_cols}
\title{Collapse timepoint and interval columns of hover info table and remove empty rows}
\usage{
collapse_cols(col1, col2, col3, col4)
}
\arguments{
\item{col1}{Column holding start dates/days from interval events.}

\item{col2}{Column holding end dates/days from interval events.}

\item{col3}{Column holding start dates/days from exp interval events.}

\item{col4}{Column holding end dates/days from exp interval events.}
}
\value{
Character vector of collapsed dates/days.
}
\description{
Collapse timepoint and interval columns of hover info table and remove empty rows
}
\keyword{internal}
