% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{color_lookup}
\alias{color_lookup}
\title{Create color lookup table}
\usage{
color_lookup(groups)
}
\arguments{
\item{groups}{Character vector of unique event types of the data to be
plotted, e.g. returned by \code{combine_data}. Defines length and element
names of the returned vector.}
}
\value{
A named vector of hexcode colors.
}
\description{
\code{color_lookup} returns a named vector of hex colors.
}
\keyword{internal}
