% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{combine_data}
\alias{combine_data}
\title{Combine subject level, event, and drug administration data frames}
\usage{
combine_data(df_list, basic_info_df, trtstart, trtend, icf_date, subjid_var)
}
\arguments{
\item{df_list}{A list of data frames as delivered by \code{set_events_intern()} and
\code{set_exp_intervals()}.}

\item{basic_info_df}{A data frame containing subject level data as extracted by
\code{set_basics()}.}

\item{trtstart}{Character name of the treatment start column as extracted
by \code{set_basics()}.}

\item{trtend}{Character name of the treatment end column as extracted by
\code{set_basics()}.}

\item{icf_date}{Character name of the informed consent date column of the
subject level dataset as extracted by \code{set_basics()}.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
A data frame in which the columns of all data frames of the df_list
are combined and with the following additional columns:
\itemize{
\item{A treatment start date column as specified through the trtstart
parameter.}
\item{A treatment end date column as specified through the trtend
parameter.}
\item{\code{earliest}: Holds the date of the earliest event per subject,
which is always the informed consent date.}
}
}
\description{
Given a list of data frames, \code{combine_data()} merges them into one data
frame and adds subject level specific columns.
}
\keyword{internal}
