% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{complete_events}
\alias{complete_events}
\title{Adapt data frame structure for plotting purposes}
\usage{
complete_events(combined_data, trt_start, trt_end)
}
\arguments{
\item{combined_data}{A data frame as provided by \code{combine_data()}.}

\item{trt_start}{Character name of the treatment start column as extracted
from \code{set_basics}.}

\item{trt_end}{Character name of the treatment end column as extracted
from \code{set_basics}.}
}
\value{
A data frame as provided via the combined_data parameter with
additional columns:
\itemize{
\item{\code{start_dy_var}: Event start days relative to study start.}
\item{\code{end_dy_var}: Event end days relative to study start.}
\item{\code{start_exp_day}: Start of time exp intervals in days relative to study
start.}
\item{\code{end_exp_day}: End of time exp intervals in days relative to study start.}
}
}
\description{
\code{complete_events()} ensures all dates necessary for the main plot, i.e. start and
and end dates/days per event are available and of respective type. If no study day
columns were determined at app configuration, the function calculates them
according to SDTM rules.
}
\keyword{internal}
