% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{create_ggdata_y}
\alias{create_ggdata_y}
\title{Generates ggplot_build table and returns only rows with data near mouse cursor}
\usage{
create_ggdata_y(p, hover)
}
\arguments{
\item{p}{ggplot object that represents the main plot.}

\item{hover}{Hover object of the main plot.}
}
\value{
A list of two data frames extracted from a call to
\code{ggplot2:::ggplot_build()}, and filtered by data points that lie within +/- 5 px
(on y-axis) from cursor position. One data frame represents interval data, the other
timepoints.
}
\description{
Generates ggplot_build table and returns only rows with data near mouse cursor
}
\keyword{internal}
