% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{create_hover_info}
\alias{create_hover_info}
\title{Create info box for hovering over the main plot of \pkg{dv.clinlines}}
\usage{
create_hover_info(
  hover,
  ggdata_y,
  initial_data,
  color_map,
  x_scale,
  y_screen_pct,
  time_range
)
}
\arguments{
\item{hover}{Hover event of the main plot.}

\item{ggdata_y}{List of data frames as returned by \code{create_ggdata_y()}.}

\item{initial_data}{Data frame returned by \code{prep_data()}.}

\item{color_map}{Named vector of hexadecimal color codes, as returned by
\code{color_lookup()}. Names represent event groups.}

\item{x_scale}{Character string containing either "date", or "day". Reveals current
x-axis setting.}

\item{y_screen_pct}{Y mouse position on window screen (between 0 and 1)}

\item{time_range}{Datetime or numeric vector, length 2. Indicates limits of
x-axis in main and detail plots.}
}
\value{
A shiny wellPanel object which displays each event type (group),
timepoint date (moment), interval dates (start, end), and further details
to the event (details) for the hovered subject of the main plot.
}
\description{
\code{create_hover_info} returns a shiny wellPanel with further event
information for the hovered-on subject in the main plot.
}
\keyword{internal}
