% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{create_main_plot}
\alias{create_main_plot}
\title{Create plot for main view of the \pkg{dv.clinlines} module}
\usage{
create_main_plot(work_data, y_sort, time_range, colors, height, subjid_var)
}
\arguments{
\item{work_data}{Data frame returned by \code{create_plot_data()}}

\item{y_sort}{Character name of selected y-axis sorting type of the main
plot. Can be either "alphanum" or "earliest".}

\item{time_range}{Datetime or numeric vector, length 2. Indicates limits of
x-axis in main and detail plots.}

\item{colors}{Named character vector of color hex codes. Must be named after
event types. Must be of the same length as there are distinct event types to
plot. Can be generated by a call to \code{color_lookup()}.}

\item{height}{Box/row height for each subject. Numeric.
Must be a single value.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
A ggplot object - the main plot of  \pkg{dv.clinlines}.
}
\description{
\code{create_main_plot} returns a ggplot object in consideration of chosen
x-axis scale, y-axis sorting type, and time range.
}
\keyword{internal}
