% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{create_plot_data}
\alias{create_plot_data}
\title{Adapt time intervals for plotting purposes}
\usage{
create_plot_data(work_data, time_range, filter_event)
}
\arguments{
\item{work_data}{Data frame returned by \code{prep_data()}.}

\item{time_range}{Datetime or numeric vector, length 2. Indicates limits of
x-axis in main and detail plots.}

\item{filter_event}{Character name of selected event type. Can be one of the
entries of \code{work_data$group}.}
}
\value{
Data frame with same structure as \code{work_data}. All start and end
variables were modified according to the chosen time range, so that no date
or day is outside this range. Events were filtered according to event type
filter settings.
}
\description{
Modifies a data frame so that interval starts and ends are not outside
time_range boundaries.
}
\keyword{internal}
