% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{create_tooltip}
\alias{create_tooltip}
\title{Create tooltip at mouse position}
\usage{
create_tooltip(hover, y_screen_pct)
}
\arguments{
\item{hover}{Hover event of the main plot.}

\item{y_screen_pct}{Y mouse position on window screen (between 0 and 1)}
}
\value{
Style property for the tooltip
}
\description{
Create tooltip at mouse position
}
\keyword{internal}
