% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dummy_data.R
\name{default_basic_info}
\alias{default_basic_info}
\title{Set default values for \code{basic_info}}
\usage{
default_basic_info()
}
\value{
A list that points to an ADSL dataset and assigns the following elements:
\itemize{
\item{\code{TRTSDT} as treatment start variable}
\item{\code{TRTEDT} as treatment end variable}
\item{\code{RFICDT} as informed consent date variable}
}
}
\description{
Set default values for \code{basic_info}
}
