% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dummy_data.R
\name{default_drug_admin}
\alias{default_drug_admin}
\title{Set default values for \code{drug_admin}}
\usage{
default_drug_admin()
}
\value{
A list that points to an ex domain dataset and assigns the following elements:
\itemize{
\item{\code{EXSTDTC} as start variable}
\item{\code{EXENDTC} as end variable}
\item{\code{EXTRT} as detail variable}
\item{\code{Drug Administration} as label}
\item{\code{EXDOSE} as exposure dosage}
\item{\code{EXDOSU} as dose unit}
}
}
\description{
Set default values for \code{drug_admin}
}
