% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dummy_data.R
\name{default_mapping}
\alias{default_mapping}
\title{Set default values for \code{mapping}}
\usage{
default_mapping()
}
\value{
A list that contains lists of events for ADSL, ADAE, and ADCM datasets. Events
defined are:
\itemize{
\item{For ADSL: Treatment Start (\code{TRTSDT} variable), Treatment End (\code{TRTEDT}
variable), and Informed Consent (\code{RFICDT} variable)}
\item{For ADAE: Adverse Events with \code{AESTDTC} (start_dt_var), \code{AEENDTC}
(end_dt_var), and \code{AEDECOD} (detail_var)}
\item{For ADCM: Concomitant Medications with \code{CMSTDTC} (start_dt_var),
\code{CMENDTC} (end_dt_var), and \code{CMDECOD} (detail_var)}
}
}
\description{
Set default values for \code{mapping}
}
