% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_local_filter.R
\name{filter_data}
\alias{filter_data}
\title{Filter data according to local filter settings}
\usage{
filter_data(data, status_filters, input_filters, filter)
}
\arguments{
\item{data}{A data frame as provided by \code{prep_data}.}

\item{status_filters}{Named boolean vector as returned
by \code{get_filter_status()}.}

\item{input_filters}{A list of inputs from local filters.}

\item{filter}{\verb{[list(list(character(1)+)) | NULL]}

A list that specifies information for local adverse events filters.
Set to \code{NULL} (default) for no filters.}
}
\value{
The received data frame filtered according to local filter settings.
}
\description{
Filter data according to local filter settings
}
\keyword{internal}
