% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{filter_nearest}
\alias{filter_nearest}
\title{Filter prepared event data by cursor position.}
\usage{
filter_nearest(initial_data, subject, rel_groups, x_range, x_scale, time_range)
}
\arguments{
\item{initial_data}{Data frame returned by \code{prep_data()}.}

\item{subject}{Character sting of one unique subject identifier.}

\item{rel_groups}{Character vector of event names as returned by \code{get_groups()}.}

\item{x_range}{Numeric vector of x values as returned by \code{calc_x_range()}.}

\item{x_scale}{Character string containing either "date", or "day". Reveals current
x-axis setting.}

\item{time_range}{Datetime or numeric vector, length 2. Indicates limits of
x-axis in main and detail plots.}
}
\value{
A data frame containing only those events of which start/end/timepoint lie
within an area determined by \code{rel_groups} and \code{x_range}.
}
\description{
Filter prepared event data by cursor position.
}
\keyword{internal}
