% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_local_filter.R
\name{get_filter_status}
\alias{get_filter_status}
\title{Get status of each available filter}
\usage{
get_filter_status(all_filters, chosen_filters)
}
\arguments{
\item{all_filters}{Character vector of all available filter names.}

\item{chosen_filters}{List of filters to be displayed.}
}
\value{
Named vector that contains for each available filter if it was
chosen by the user to be displayed (TRUE) or not (FALSE).
}
\description{
Checks for each available filter if it was chosen to be displayed in the
user interface of \pkg{dv.clinlines}.
}
\keyword{internal}
