% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_main_view.R
\name{get_groups}
\alias{get_groups}
\title{Extract group names of events near cursor position}
\usage{
get_groups(ggdata_y, color_map, x_range)
}
\arguments{
\item{ggdata_y}{List of data frames as returned by \code{create_ggdata_y()}.}

\item{color_map}{Named vector of hexadecimal color codes, as returned by
\code{color_lookup()}. Names represent event groups.}

\item{x_range}{Vector of start and end values on x-axis level as provided by
\code{calc_x_range}.}
}
\value{
Character vector of event names.
}
\description{
Extract group names of events near cursor position
}
\keyword{internal}
