% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/local_filter_setup.R
\name{local_filters}
\alias{local_filters}
\title{UI storage of available local adverse event filters}
\usage{
local_filters(ns)
}
\arguments{
\item{ns}{Namespace (character).}
}
\value{
A list UI elements that represents all available local adverse event filters
}
\description{
UI storage of available local adverse event filters
}
\keyword{internal}
