% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_clinical_timelines.R
\name{mock_clinical_timelines_app}
\alias{mock_clinical_timelines_app}
\title{Run clinlines mock}
\usage{
mock_clinical_timelines_app()
}
\description{
\code{mock_clinical_timelines_app()} runs the \pkg{dv.clinlines} module
with dummy data. Local adverse event filters included.
}
