% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_clinical_timelines.R
\name{mock_clinical_timelines_server}
\alias{mock_clinical_timelines_server}
\title{Mock server for clinlines module}
\usage{
mock_clinical_timelines_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny.}
}
\description{
Uses adsl, adae, and adcm data of roche dataset.
}
\keyword{internal}
