% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_clinical_timelines.R
\name{mock_clinical_timelines_UI}
\alias{mock_clinical_timelines_UI}
\title{Mock UI for clinlines module}
\usage{
mock_clinical_timelines_UI(id = NULL)
}
\arguments{
\item{id}{A unique identifier.}
}
\value{
Shiny UI for the clinlines module.
}
\description{
Mock UI for clinlines module
}
\keyword{internal}
