% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_local_filter.R
\name{mock_local_filter_app}
\alias{mock_local_filter_app}
\title{Run local filters mock}
\usage{
mock_local_filter_app()
}
\description{
\code{mock_local_filter_app()} runs the local filter module with
dummy data.
}
\keyword{internal}
