% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_local_filter.R
\name{mock_local_filter_server}
\alias{mock_local_filter_server}
\title{Mock server for the local filter module}
\usage{
mock_local_filter_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny.}
}
\description{
Uses adsl, adae, and adcm datasets of roche data.
}
\keyword{internal}
