% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_local_filter.R
\name{mock_local_filter_UI}
\alias{mock_local_filter_UI}
\title{Mock UI for the local filter module}
\usage{
mock_local_filter_UI(request)
}
\arguments{
\item{request}{Required parameter for bookmarking.}
}
\value{
Shiny UI for the local filters module.
}
\description{
Mock UI for the local filter module
}
\keyword{internal}
