% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_main_view.R
\name{mock_main_view_app}
\alias{mock_main_view_app}
\title{Run main view mock app}
\usage{
mock_main_view_app()
}
\description{
\code{mock_main_view_app()} runs the main view module with dummy data.
}
\keyword{internal}
