% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_main_view.R
\name{mock_main_view_server}
\alias{mock_main_view_server}
\title{Mock server for main view module of clinlines}
\usage{
mock_main_view_server(input, output, session)
}
\arguments{
\item{input, output, session}{Internal parameters for shiny.}
}
\description{
Uses adsl, adae, and adcm datasets of roche data.
}
\keyword{internal}
