% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_main_view.R
\name{mock_main_view_UI}
\alias{mock_main_view_UI}
\title{Mock UI for main view module of clinlines}
\usage{
mock_main_view_UI()
}
\value{
Shiny UI for the main view module.
}
\description{
Mock UI for main view module of clinlines
}
\keyword{internal}
