% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mock_with_mm.R
\name{mock_with_mm_app}
\alias{mock_with_mm_app}
\title{Run an example for Clinical Timelines integrated in the module manager}
\usage{
mock_with_mm_app()
}
\description{
Launches an example app that shows a Clinical Timelines module integrated in the
module manager surface. Displays data from the \pkg{pharmaverseadam} package.
}
\keyword{internal}
