% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clinical_timelines.R
\name{mod_clinical_timelines_UI}
\alias{mod_clinical_timelines_UI}
\title{Create user interface for \pkg{dv.clinlines} module}
\usage{
mod_clinical_timelines_UI(
  module_id,
  filter_list = NULL,
  x_param = "day",
  boxheight_val = 60
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A unique ID string to create a namespace. Must match the ID of
\code{mod_clinical_timelines_server()}.}

\item{filter_list}{\verb{[list(character(1)+) | NULL]}

List of filter names that indicate which local filters to
display.}

\item{x_param}{\verb{[character(1) | NULL]}

Either "date" or "day" defining if the x axis shows the date or
study days as initial setting at app launch. Defaults to "day".}

\item{boxheight_val}{\verb{[integer(1) | NULL]}

A value between 30 and 150 defining the initial height of
the individual timeline plot boxes at app launch. Defaults to 60.}
}
\value{
A shiny UI \code{tagList} that contains a \code{sidebarLayout} object.
}
\description{
\code{mod_clinical_timelines_UI} acts as central for managing user interfaces of
main view and local filter modules.
}
\seealso{
\code{\link[=mod_clinical_timelines]{mod_clinical_timelines()}} and \code{\link[=mod_clinical_timelines_server]{mod_clinical_timelines_server()}}
}
