% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_local_filter.R
\name{mod_local_filter_server}
\alias{mod_local_filter_server}
\title{Create server for local filter shiny module of \pkg{dv.clinlines}}
\usage{
mod_local_filter_server(module_id, filter, joined_data, changed)
}
\arguments{
\item{module_id}{A unique ID string to create a namespace. Must match the ID of
\code{mod_local_filter_UI()}.}

\item{filter}{\verb{[list(list(character(1)+)) | NULL]}

A list that specifies information for local adverse events filters.
Set to \code{NULL} (default) for no filters.}

\item{joined_data}{A metaReactive dataset as provided by \code{prep_data()}.}

\item{changed}{A reactive whose actualization indicates if the filters should be reset to default values.}
}
\value{
A data frame filtered by local adverse event filters.
}
\description{
Create server for local filter shiny module of \pkg{dv.clinlines}
}
\keyword{internal}
