% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_local_filter.R
\name{mod_local_filter_UI}
\alias{mod_local_filter_UI}
\title{Create user interface for the local filter shiny module of \pkg{dv.clinlines}}
\usage{
mod_local_filter_UI(module_id, filter_list = NULL)
}
\arguments{
\item{module_id}{A unique ID string to create a namespace. Must match the ID of
\code{mod_local_filter_server()}.}

\item{filter_list}{List of filter names that indicate which local filters to
display.}
}
\value{
A shiny \code{uiOutput} element.
}
\description{
Create user interface for the local filter shiny module of \pkg{dv.clinlines}
}
\keyword{internal}
