% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_main_view.R
\name{mod_main_view_UI}
\alias{mod_main_view_UI}
\title{Create user interface for the main view module of \pkg{dv.clinlines}}
\usage{
mod_main_view_UI(module_id, x_param = "day", boxheight_val = 60)
}
\arguments{
\item{module_id}{A unique ID string to create a namespace. Must match the ID of
\code{mod_main_view_server()}.}

\item{x_param}{Either "date" or "day" defining if the x axis shows the date or
study days as initial setting at app launch. Defaults to "day".}

\item{boxheight_val}{A value between 30 and 150 defining the initial height of
the individual timeline plot boxes at app launch. Defaults to 60.}
}
\value{
A list containing two entries:
\itemize{
\item{\code{sidebar}: A \code{tagList} of shiny input elements.}
\item{\code{main_panel}: A \code{div} element containing shiny UI elements.}
}
}
\description{
Create user interface for the main view module of \pkg{dv.clinlines}
}
\keyword{internal}
