% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_dummy_data.R
\name{prep_dummy_data}
\alias{prep_dummy_data}
\title{Prepare dummy data}
\usage{
prep_dummy_data(n = 200)
}
\arguments{
\item{n}{Number of rows to select from the adsl dataset. The first n rows will be
taken. Used to reduce runtime during development.}
}
\value{
A list of three data frames.
}
\description{
Modifiy pharmaverseadam's adsl, adae, adcm, and exp dummy data for easy use within
\pkg{dv.clinlines}.
}
\keyword{internal}
