% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{set_basics}
\alias{set_basics}
\title{Set information about subject level analysis data (e.g. adsl, dm, ...)}
\usage{
set_basics(data_list, basic_info = default_basic_info(), subjid_var)
}
\arguments{
\item{data_list}{A named list of data frames that provides a subject level dataset
and further datasets with event specific data. Usually obtained from module manager.}

\item{basic_info}{\verb{[list(character(1)+)]}

A list of four elements: \code{subject_level_dataset_name}, \code{trt_start_var},
\code{trt_end_var}, and \code{icf_date_var}. Assigns the name
of a subject level dataset and column names of treatment start and end, and informed
consent variables.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
A list containing the following elements:
\itemize{
\item{\code{data}: A data frame that includes subject level data as
provided in \code{data_list} with name for the subject identifier column fixed to
\code{subjid_var}.}
\item{\code{trt_start}: Character name of the variable that contains
treatment start dates as provided in \code{basic_info}.}
\item{\code{trt_end}: Character name of the variable that contains
treatment end dates as provided in \code{basic_info}.}
\item{\code{icf_date}: Character name of the variable that contains
informed consent dates as provided in \code{basic_info}.}
}
}
\description{
Set information about subject level analysis data (e.g. adsl, dm, ...)
}
\keyword{internal}
