% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{set_events_intern}
\alias{set_events_intern}
\title{Set and prepare interval data}
\usage{
set_events_intern(data_list, mapping = default_mapping(), subjid_var)
}
\arguments{
\item{data_list}{A named list of data frames that provides a subject level dataset
and further datasets with event specific data. Usually obtained from module manager.}

\item{mapping}{\verb{[list(list(list(character(1)+)))]}

A list of lists. It serves as instruction on which event to take from
which data domain / dataset, and further from which variables to take the start and
end values, etc., that will be plotted as clinical timelines. Elements need to follow
a certain structure that is described in the Details section below.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
A data frame including the following columns:
\itemize{
\item{\code{subject_id}: A unique subject identifier column. Its name is specified by
\code{subjid_var}.}
\item{\code{start_dt_var}, \code{end_dt_var}: Start and end dates of the event.}
\item{\code{detail_var}: Further information to the events.}
\item{\code{set}: Name of the dataset the data origins from.}
\item{\code{set_id}: Row ID's of the related dataset.}
\item{\code{group}: Label for the event types.}
\item{\code{arrow_right}: Flag that indicates whether the event is ongoing/ended
after the specified time range or not.}
}
}
\description{
Gathers data needed for the plot and transforms them
to a consistent structure.
}
\details{
Please refer to the Details section of \code{mod_clinical_timelines()} for further
instructions on how to define a proper mapping.
}
\keyword{internal}
