% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{set_exp_intervals}
\alias{set_exp_intervals}
\title{Set and prepare interval for exp data}
\usage{
set_exp_intervals(data_list, mapping = default_drug_admin(), subjid_var)
}
\arguments{
\item{data_list}{A named list of data frames that provides a subject level dataset
and further datasets with event specific data. Usually obtained from module manager.}

\item{mapping}{A list of named character strings that describes which variables to
use to display drug administration events.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
A data frame including the following columns:
\itemize{
\item{\code{subject_id}: A unique subject identifier column. Its name is specified by
\code{subjid_var}.}
\item{\code{start_exp}, \code{end_exp}: Start and end dates of the event.}
\item{\code{detail_var}: Information about drug name and dosage.}
\item{\code{set}: Name of the dataset the data origins from.}
\item{\code{set_id}: Row ID's of the related dataset.}
\item{\code{dose_var}: Indicates whether the dose of the treatment has increased,
decreased or stayed the same compared to the last dose at the same subject.}
\item{\code{group}: Label for the event types.}
}
}
\description{
Gathers data needed for the plotting of drug administration periods and transforms them
to a consistent structure.
}
\details{
The \code{mapping} list must contain the following elements:
\itemize{
\item{\code{name}: Character name of the dataset that holds drug administration data
(e.g. ex domain), as it is called in the datalist that is provided to the
\pkg{modulemanager}.}
\item{\code{start_var}: Character name of the variable that contains the start dates
(e.g. exposure start dates) which must be present in the dataset mentioned in the
\code{name} element.}
\item{\code{end_var}: Character name of the variable that contains the end dates
(e.g. exposure end dates) which must be present in the dataset mentioned in the
\code{name} element.}
\item{\code{detail_var}: Character name of the variable that contains the treatment
information. Must exist in the dataset mentioned in the \code{name} element.}
\item{\code{label}: Free-text character label for the drug administration event.}
\item{\code{dose_var}: Character name of the variable that contains the dosis level
information. Must exist in the dataset mentioned in the \code{name} element.}
\item{\code{dose_unit_var}: Character name of the variable that contains the dosis
unit. Must exist in the dataset mentioned in the \code{name} element.}
}
}
\keyword{internal}
