% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_prep.R
\name{set_filter_dataset}
\alias{set_filter_dataset}
\title{Create a dataset for local adverse event filters}
\usage{
set_filter_dataset(filter, data_list, mapping, subjid_var)
}
\arguments{
\item{filter}{\verb{[list(list(character(1)+)) | NULL]}

A list that specifies information for local adverse events filters.
Set to \code{NULL} (default) for no filters.}

\item{data_list}{A named list of data frames that provides a subject level dataset
and further datasets with event specific data. Usually obtained from module manager.}

\item{mapping}{\verb{[list(list(list(character(1)+)))]}

A list of lists. It serves as instruction on which event to take from
which data domain / dataset, and further from which variables to take the start and
end values, etc., that will be plotted as clinical timelines. Elements need to follow
a certain structure that is described in the Details section below.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}
}
\value{
A subset of the adverse event dataset containing all columns needed
for local adverse event filters.
}
\description{
Create a dataset for local adverse event filters
}
\details{
The list provided to \code{mapping} must follow a strict hierarchy. It must contain
one entry per dataset/domain that serves as basis for the events. These entries need to
be named according to the names of the datalist that is provided to the
\pkg{modulemanager}. The entries by oneself must again be lists. \cr
These second level lists contain the variable names that are needed to plot the events,
gathered in yet another lists, which are named according to the labels that shall be
assigned to each event, and that contain the following elements each:
\itemize{
\item{\code{start_dt_var}: Character name of the variable that contains
either the event start dates (for interval events) or merely timepoints (e.g.
milestones). The variable name must be present in the dataset under which the
event is listed.}
\item{\code{end_dt_var}: Character name of the variable that contains
the event end dates. Needs to be provided for interval events, but set to
\code{NULL} for timepoints. The variable name must be present in the dataset under
which the event is listed.}
\item{\code{start_dy_var}: Similar to \code{start_dt_var}, but refers to the study
relative days (instead of dates). The variable name must be present in the dataset
under which the event is listed. Can be set to \code{NULL} to let the module
calculate the study days according to SDTM standard rules.}
\item{\code{end_dy_var}: Similar to \code{end_dt_var}, but refers to the study
relative days (instead of dates). The variable name must be present in the dataset
under which the event is listed. Can be set to \code{NULL} to let the module
calculate the study days according to SDTM standard rules.}
\item{\code{detail_var}: Character name of the variable that contains further
descriptive information that shall be displayed for the event. Can be set to
\code{NULL} for no further information.}
}

The structure of the \code{mapping} parameters for one single event is mentioned below.
It is possible to define multiple events for one dataset, and multiple datasets in
the mapping list.

\code{mapping = list(}\cr
\verb{  }\code{<data name> = list}(\cr
\verb{  }\verb{  }\code{<event label> = list}(\cr
\verb{  }\verb{  }\verb{  }\code{start_dt_var = <variable name>,}\cr
\verb{  }\verb{  }\verb{  }\code{end_dt_var = <variable name or NULL>,}\cr
\verb{  }\verb{  }\verb{  }\code{start_dy_var = <variable name or NULL>,}\cr
\verb{  }\verb{  }\verb{  }\code{end_dy_var = <variable name or NULL>,}\cr
\verb{  }\verb{  }\verb{  }\code{detail_var = <variable name or NULL>}\cr
\verb{  }\verb{  }\code{)}\cr
\verb{  }\code{)}\cr
\code{)}
}
\keyword{internal}
