% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server_functions_local_filter.R
\name{set_pts}
\alias{set_pts}
\title{Set choices for Preferred Term filter}
\usage{
set_pts(status_filters, soc, prepped_data, filter)
}
\arguments{
\item{status_filters}{Named boolean vector as returned
by \code{get_filter_status()}.}

\item{soc}{Character vector of System Organ Classes.}

\item{prepped_data}{Data frame as returned by \code{prep_data()}.}

\item{filter}{\verb{[list(list(character(1)+)) | NULL]}

A list that specifies information for local adverse events filters.
Set to \code{NULL} (default) for no filters.}
}
\value{
Character vector of PT's.
}
\description{
Set choices for Preferred Term filter
}
\keyword{internal}
