% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mod_clinical_timelines.R
\name{mod_clinical_timelines}
\alias{mod_clinical_timelines}
\title{Wrapper that serves as interface for defining an application in the \pkg{modulemanager}}
\usage{
mod_clinical_timelines(
  module_id,
  basic_info = default_basic_info(),
  mapping = default_mapping(),
  drug_admin = default_drug_admin(),
  filter = NULL,
  subjid_var = "USUBJID",
  default_plot_settings = list(x_param = "day", start_day = NULL, boxheight_val = 60),
  color_palette = NULL,
  ms = 1000,
  receiver_id = NULL
)
}
\arguments{
\item{module_id}{\verb{[character(1)]}

A character string that serves as unique identifier for the module.}

\item{basic_info}{\verb{[list(character(1)+)]}

A list of four elements: \code{subject_level_dataset_name}, \code{trt_start_var},
\code{trt_end_var}, and \code{icf_date_var}. Assigns the name
of a subject level dataset and column names of treatment start and end, and informed
consent variables.}

\item{mapping}{\verb{[list(list(list(character(1)+)))]}

A list of lists. It serves as instruction on which event to take from
which data domain / dataset, and further from which variables to take the start and
end values, etc., that will be plotted as clinical timelines. Elements need to follow
a certain structure that is described in the Details section below.}

\item{drug_admin}{\verb{[list(character(1)+) | NULL]}

A list of named character strings that describes which variables to
use to display drug administration events. Set it to NULL to deactivate them.}

\item{filter}{\verb{[list(list(character(1)+)) | NULL]}

A list that specifies information for local adverse events filters.
Set to \code{NULL} (default) for no filters.}

\item{subjid_var}{\verb{[character(1)]}

Character name of the unique subject identifier column in
all datasets (default is USUBJID). Must be a single value.}

\item{default_plot_settings}{\verb{[list(character(1), integer(1), integer(1)) | NULL]}

A named list containing three elements: \code{x_param} contains
either "date" or "day" defining if the x axis shows the date or
study days as default setting at app launch (defaults to "day"), \code{start_day}
contains an integer indicating the lower x-axis limit in case of study day display
(defaults to NULL, using the day of the earliest event to be displayed),
\code{boxheight_val} contains a value between 30 and 150 defining the initial height of
the individual timeline plot boxes at app launch (defaults to 60).}

\item{color_palette}{\verb{[character(1+) | NULL]}

A named vector that specifies the colors for drawing events or intervals.
Each name in the vector should correspond to an entry in the legend.
If \code{NULL} (default), the default color palette is used.}

\item{ms}{\verb{[numeric(1)]}

Single numeric value indicating how many milliseconds to wait before the plot
re-renders. Defaults to 1000.}

\item{receiver_id}{\verb{[character(1) | NULL]}

Character string defining the ID of the module to which to send a subject ID. The
module must exist in the module list. The default is NULL which disables communication.}
}
\value{
A list containing the following elements to be used by the
\pkg{modulemanager}:
\itemize{
\item{\code{ui}: A UI function of the \pkg{dv.clinlines} module.}
\item{\code{server}: A server function of the \pkg{dv.clinlines} module.}
\item{\code{module_id}: A unique identifier.}
}
}
\description{
Wrapper that serves as interface for defining an application in the \pkg{modulemanager}
}
\details{
The \code{basic_info} list must contain the following elements:
\itemize{
\item{\code{subject_level_dataset_name}: Character name of the subject level analysis dataset
(e.g. "adsl", "dm") as it is called in the datalist that is provided to the
\pkg{modulemanager}.}
\item{\code{trt_start_var}: Character name of the variable that contains
treatment start dates which must be present in the dataset mentioned in the
\code{data} element.}
\item{\code{trt_end_var}: Character name of the variable that contains
treatment end dates which must be present in the dataset mentioned in the
\code{data} element.}
\item{\code{icf_date_var}: Character name of the variable that contains
informed consent dates which must be present in the dataset mentioned in the
\code{data} element.}
}

\cr

The list provided to \code{mapping} must follow a strict hierarchy. It must contain
one entry per dataset/domain that serves as basis for the events. These entries need to
be named according to the names of the datalist that is provided to the
\pkg{modulemanager}. The entries by oneself must again be lists. \cr
These second level lists contain the variable names that are needed to plot the events,
gathered in yet another lists, which are named according to the labels that shall be
assigned to each event, and that contain the following elements each:
\itemize{
\item{\code{start_dt_var}: Character name of the variable that contains
either the event start dates (for interval events) or merely timepoints (e.g.
milestones). The variable name must be present in the dataset under which the
event is listed.}
\item{\code{end_dt_var}: Character name of the variable that contains
the event end dates. Needs to be provided for interval events, but set to
\code{NULL} for timepoints. The variable name must be present in the dataset under
which the event is listed.}
\item{\code{start_dy_var}: Similar to \code{start_dt_var}, but refers to the study
relative days (instead of dates). The variable name must be present in the dataset
under which the event is listed. Can be set to \code{NULL} to let the module
calculate the study days according to SDTM standard rules.}
\item{\code{end_dy_var}: Similar to \code{end_dt_var}, but refers to the study
relative days (instead of dates). The variable name must be present in the dataset
under which the event is listed. Can be set to \code{NULL} to let the module
calculate the study days according to SDTM standard rules.}
\item{\code{detail_var}: Character name of the variable that contains further
descriptive information that shall be displayed for the event. Can be set to
\code{NULL} for no further information.}
}

The structure of the \code{mapping} parameters for one single event is mentioned below.
It is possible to define multiple events for one dataset, and multiple datasets in
the mapping list.

\code{mapping = list(}\cr
\verb{  }\code{<data name> = list}(\cr
\verb{  }\verb{  }\code{<event label> = list}(\cr
\verb{  }\verb{  }\verb{  }\code{start_dt_var = <variable name>,}\cr
\verb{  }\verb{  }\verb{  }\code{end_dt_var = <variable name or NULL>,}\cr
\verb{  }\verb{  }\verb{  }\code{start_dy_var = <variable name or NULL>,}\cr
\verb{  }\verb{  }\verb{  }\code{end_dy_var = <variable name or NULL>,}\cr
\verb{  }\verb{  }\verb{  }\code{detail_var = <variable name or NULL>}\cr
\verb{  }\verb{  }\code{)}\cr
\verb{  }\code{)}\cr
\code{)}

\cr

If not NULL, the \code{drug_admin} list must contain the following elements (with the
exception of \code{show_dose_info} being optional):
\itemize{
\item{\code{dataset_name}: Character name of the dataset that holds drug administration data
(e.g. ex domain), as it is called in the datalist that is provided to the
\pkg{modulemanager}.}
\item{\code{trt_var}: Character name of the variable that contains the
treatment name. Must exist in the dataset mentioned in the
\code{dataset_name} element.}
\item{\code{start_var}: Character name of the variable that contains the start dates
(e.g. exposure start dates). Must exist in the dataset mentioned in the
\code{dataset_name} element.}
\item{\code{end_var}: Character name of the variable that contains the end dates
(e.g. exposure end dates). Must exist in the dataset mentioned in the
\code{dataset_name} element.}
\item{\code{detail_var}: Character name of the variable that contains the treatment
information to appear in the Details column of the hover box. Must exist in the dataset
mentioned in the \code{dataset_name} element.}
\item{\code{label}: Free-text character label for the drug administration event.}
\item{\code{dose_var}: Character name of the variable that contains the dose level
information. Must exist in the dataset mentioned in the \code{dataset_name} element.}
\item{\code{dose_unit_var}: Character name of the variable that contains the dose
unit. Must exist in the dataset mentioned in the \code{dataset_name} element.}
\item{\code{show_dose_info}: Logical indicating whether or not to show dose level and
unit in the Details column of the hover box.}
}

\cr

If not \code{NULL}, \code{filter} defines local filters.
So far, the following filters for adverse events are available:
\itemize{
\item{\code{serious_ae_var}: Filter for serious adverse events.}
\item{\code{soc_var}: Filter for system organ classes.}
\item{\code{pref_term_var}: Filter for preferred terms.}
\item{\code{drug_rel_ae_var}: Filter for drug related adverse events.}
}
The \code{filter} parameter must be a list that contains yet another list for adverse
events filters, that is named \code{ae_filter}, and that holds the following elements:
\itemize{
\item{\code{dataset_name}: Character name of the adverse events dataset. Must be
available in in the datalist that is provided to the \pkg{modulemanager}.}
\item{\code{label}: Character value which is exactly the same as the name for the
adverse events event defined in \code{mapping}.}
\item{\code{serious_ae_var}: Character name of the adverse events variable that contains
serious adverse events flags (Y/N).}
\item{\code{soc_var}: Character name of the adverse events variable that contains
system organ classes.}
\item{\code{pref_term_var}: Character name of the adverse events variable that contains
preferred terms.}
\item{\code{drug_rel_ae_var}: Character name of the adverse events variable that contains
drug related (causality) flags (Y/N).}
}
}
\seealso{
\code{\link[=mod_clinical_timelines_UI]{mod_clinical_timelines_UI()}}, \code{\link[=mod_clinical_timelines_server]{mod_clinical_timelines_server()}},
\code{\link[=set_basic_info]{set_basic_info()}}, \code{\link[=set_event]{set_event()}}, \code{\link[=set_drug_admin]{set_drug_admin()}}
}
