% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{set_basic_info}
\alias{set_basic_info}
\title{Create a list to be assigned to the \code{basic_info} parameter}
\usage{
set_basic_info(data, trt_start, trt_end, icf_date)
}
\arguments{
\item{data}{Character name of the subject level analysis dataset (e.g. "adsl", "dm")
as it is called in the \code{data_list} parameter.}

\item{trt_start}{Character name of the variable that contains treatment start dates.
Must be present in the data frame mentioned in the \code{data} element.}

\item{trt_end}{Character name of the variable that contains treatment end dates.
Must be present in the data frame mentioned in the \code{data} element.}

\item{icf_date}{Character name of the variable that contains informed consent dates.
Must be present in the data frame mentioned in the \code{data} element.}
}
\value{
A list that could directly be used as input for the \code{basic_info} parameter
of \code{mod_clinical_timelines()} and \code{mod_clinical_timelines_server()}.
}
\description{
Helper function to support configuration of a Clinical Timelines module.
}
